RSA/requirements.txt
Flask>=2.1.2
Celery>=5.2.6
redis>=4.2.2
Pillow>=9.0.0
numpy>=1.22.3
flask-cors>=3.1.1
werkzeug>=2.1.2
